Added getAllAvailableSpecies() method - support for FASTA
I added one info endpoints in order to get a list of all available
species. Now a fasta content-type response is returned as a Bio.Seqrecord
object. I have to decide the returning strategy: I have to read
the response with an appropriate module or I have to retorn the
raw data from REST server?