First Rest CGI script
I've completed a script in order to do GCI request via rest by accepting
a file in input. There are some minor changes to do: SNPs are always in
positive strands. The allele I read from the input file is the variant, I
don't have the reference. So I've to change the input file (I need to use VCF)
and I have to deal with reference position. Could I assing a default value of "N"?
I've to control the VEP ensembl script