First attempt to retrive reference sequence from EnsEMBL
I wrote a module to retrive ref sequence for snpchimp variants
and I've integrated it into SNPchiMpVep.py. The time for view
cgi-page depends on how many request I do with REST service and
how long they are. If variant is the same of reference sequence,
result will filtered out.